pyproject.toml
[project]
name = "hermes"
version = "0.1.0"
description = "Humanlike Email Responses for Magically Empathic Sales"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "poethepoet>=0.34.0",
    "pandas>=2.2.0",
    "openai>=1.0.0",
    "gspread>=6.0.0",
    "gspread-dataframe>=4.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "pydantic>=2.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "nest-asyncio>=1.6.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.4.0",
    "langsmith>=0.3.0",
    "pydrive2>=1.21.0",
    "oauth2client>=4.0.0",
    "langchain-core>=0.3.60",
    "chromadb>=1.0.9",
    "PyYAML>=6.0",
    "sentence-transformers>=4.0.0",
    "openevals>=0.1.0",
    "langgraph-cli[inmem]>=0.2.10",
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.3.250308",
    "thefuzz>=0.22.1",
]

[tool.poe.tasks]
start = "uv poe load_env && python -m src.hermes.main"
lint = "ruff check ."
format = "ruff format ."
load_env = "set -a; source .env; set +a"
# Add --limit N as needed, e.g., poe test-flow -- --limit 5
test-flow = "python -m tests.integration.test_agent_flow"
# Default evaluation task - uses the more efficient master evaluator
evaluate = "python -m tools.evaluate.main --dataset-id b43d1e52-72ae-47c8-9a49-55198eb6dd87 --auto-upload"
# Legacy evaluation task - kept for backward compatibility
evaluate-legacy = "python -m tools.evaluate_custom --auto-upload --dataset-id b43d1e52-72ae-47c8-9a49-55198eb6dd87 --result-dir experiments"
dev = "langgraph dev"

[dependency-groups]
dev = [
    "basedpyright>=1.29.1",
    "mypy>=1.15.0",
    "ruff>=0.11.10",
]

[tool.ruff]
exclude = [
    "notebooks",
    "docs",
]

line-length = 120
indent-width = 4

[tool.ruff.lint]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = 88
